Address workflow authoring SDK design feedback (points 5-8)
- Rename positioning to fluent builder; stop claiming full compile-time typing
- Add text()/WorkflowComposeText for explicit prompt/reference composition
- Accept builders at applyDefinition wire boundary (TS Workflow subclass, Python _workflow)
- Move golden fixture parsers/canonicalizers to language-specific test modules
- Remove test machinery from public SDK exports; keep Rust helpers doc-hidden for integration tests
- Drop parallel applyWorkflowDefinition/apply_workflow_definition from public TS/Python exports
